Corrected class main, moved class MyStack to new package collections and removed file Assignments.iml from remote repository
