+++ src/main/java/fundamentals/collections/MyStack.java
@@ -1,4 +1,4 @@
-package fundamentals;
+package fundamentals.collections;
 
 import java.util.ArrayList;
 import java.util.EmptyStackException;
